requirements.txt
conllu==4.5.3
hydra-core==1.3.2
networkx==3.2.1
transformers==4.40.2
jax
dm-haiku
optax
torch